feat: scale modules by `unpackedSize`, closes #132 (#211)
In this PR:
* Scale module `fontsize` based on `unpackedSize`
* Show `unpackedSize` in the module pane (I've just slapped this in
there for now so you can click on a module to see what it's actual size
is. I'll figure out a better UI for this later.)

The scaling factor varies as `log10(unpackedSize)`, so roughly speaking...
* 0-10KB -> 11pt
* 100KB -> 22pt
* 1MB -> 33pt
* 10MB -> 44pt... and so on

NOTE: `unpackedSize` data isn't available on module versions published prior to Feb 2018.