fix: switch theme colors from hsl to oklch
Found myself fiddling with colors in the colorizer legends while working
on #326, and ended up going all-in on the `oklch`. This is admittedly a
bit of an experiment, an excuse to learn about the `oklch` colorspace.
It's [kinda weird](https://lch.oklch.com/#0.576,78.78,216.36,100), but
has the advantage of providing colors with a consistent brightness when
you vary the H(ue) value.

Also fixing up a few odds and ends that I turned up along the way ...
* NPMSIO colorizer legend uses CSS gradient with `scoreColor()` computed
endpoints
* Use `color-mix` for color interpolation
* Tweaks to colors and styling in release timeline chart
* Wider color variation in treemap

@fregante: The tl;dr: is "a bit more color"...

Before:
![CleanShot 2025-06-16 at 10 29
17](https://github.com/user-attachments/assets/44b4c0f1-1d14-4369-b336-a02dab6cc74f)

After:
![CleanShot 2025-06-16 at 10 30
19](https://github.com/user-attachments/assets/ca11bf8a-99a6-4190-bbf4-f3b8d732fd9c)