feat: support custom registry endpoints (#332)
Adds a field to the Start tab for [optionally] specifying a custom
registry. This will allow organizations that host their own NPM
registries (e.g. with [Verdaccio](https://verdaccio.org/)) to run
registry requests through those servers rather than registry.npmjs.org.

And for a bit of extra sugar, adds a registry health check indicator
(based on whether or not a request to "/_" returns a response).

The `registry` state is stored as a search (not hash) param in the URL,
since this field affects the actual state available to the graph, not
just how it's viewed.

> [!NOTE]
> The configured registry is responsible for resolving all module
requests. I.e. if the registry is set to a private server, that server
should proxy requests for public NPM modules to `registry.npmjs.org` (or
other public mirror).

Closes #19